Final version of the algorithm
In the previous versions of the algorithm, u(dot in unicode__0t[:, :, 1] was being updated to the same value as u(dot in unicode)_1t at the end of each loop. But in the while statement in each loop, we needed to check for the convergence between u(dot in unicode)_0t and u(dot in unicode)_1t. 0t is the original guess at the beginning of the loop and 1t is the updated guess from equation 17 in the Calendino paper. So there was no way to check for the previous guess as u(dot in unicode)_0t was being updated to be the same as u(dot in unicode)_1t which should be the updated value. Here I added u(dot in unicode)_0t_old to keep track of the value of u(dot in unicode)_0t at the beginning of every loop. The code for the while loop is finally as it is supposed to be but the convergence criteria might have to be adjusted. The data for the initial values of the given varibles will definitely have to be adjusted because the data doesn't give a numercial result with the present random values. It will need to follow actual economic data.