Final version of algorithm using random data
In the previous versions of the algorithm, udot_0t[:, :, 1] was being updated to the same value as udot_1t at the end of each loop. But in the while statement in each loop, we needed to check for the convergence between udot_0t and udot_1t. 0t is the original guess at the beginning of the loop and 1t is the updated guess from equation 17 in the Calendino paper. So there was no way to check for the previous guess as udot_0t was being updated to be the same as udot_1t which should be the updated value. Here I added udot_0t_old to keep track of the value of udot_0t at the beginning of every loop. The code for the while loop is finally as it is supposed to be but the convergence criteria might have to be adjusted. The data for the initial values of the given varibles will definitely have to be adjusted because the data doesn't give a numercial result with the present random values. It will need to follow actual economic data.